agent_qms/vlm/requirements.txt
# ⚠️ DEPRECATED: This file is kept for reference only.
# All dependencies are now managed via uv in the root pyproject.toml.
#
# To install dependencies, use from the project root:
#   uv sync
#
# To run VLM commands, use:
#   uv run python -m agent_qms.vlm.cli.analyze_image_defects ...
#
# See README.md for updated installation and usage instructions.

# Core Dependencies
pydantic>=2.8.0
Pillow>=10.1.0
httpx>=0.27.0
Jinja2>=3.1.0
PyYAML>=6.0.1
python-dotenv>=1.0.1

# Optional Backend Dependencies
openai>=1.109.1  # For OpenRouter backend

# Development Dependencies (optional)
pytest>=8.4.2
ruff>=0.13.2
mypy>=1.0.0
